Fixed Spacial Pyramid Matching. Works like a charm.
The mathematical explanation in the paper was particularly
counterintuitive. The way I append all the histograms values in a single
list for comparison is an optimization over a sum of dot products, which
also simplifies the concept.

MAXRES seems doesn't need to be high, anyway that log2 is the upper bound.