Minor fixup and discussion
Got some ideas by playing around with parameters. Seems that spacial
pyramid matching is very reliable on high number of trainings, and does
not need untraining. Also, hik works well on higher numbers.

It is though necessary to make the centroids more... appetible for the
training. I tried with biasing the random initialization, but a better
way seems to be to have the centroid randomly drawn from the image's
distribution.